Add useNegativeMargin property to storycard json
On Netflix, some of the StoryCard images use a negative margin-top to
reduce white space between the image and text above. So, I've added a
a `useNegativeMargin` property to storycard-content-json. This can then
be passed to the styled-component as a prop to determine wether to apply
either `margin: -10% auto` 0 or `margin: 0 auto`.

I've also given the image a `postion: relative` and `z-index: -1` so
that the text sits above the image.