Move form & input components to GeneralFrom
I want to be able to reuse the form components used in RegForm.
So, I moved them to their own file and employed a compound components
method of default exporting a container component, ie `GeneralForm`, and
attaching related components as properties.

So, when you import GeneralFrom, all related components come along with
it and can be accessed using dot notation, eg `<GeneralForm.Input />`.