Add NavSliderBackground for opacity decrease
On NetFlix, the background opacity decreases when you open the nav
panel. I originally achieved this by passing down a prop to
MainContainer to decrease opacity. This not only creates a new stacking
context but also causes a rerender of all the genreContent rows.

I noticed that Netflix simply use a div with a transparent background
color that sits behind the Nav panel. When the panel opens its display
property is switched from `none` to `block`. So, I've implemented this
method in my clone.