RF to place currentUser into localStorage
It seems that when a user types in a URL (as opposed to redirecting with
useNavigate), the entire React DOM tree is rebuilt! This mean that
initally there is no user stored in the Context state until firebase can
be called again. This was causing me problems with protected routes, as
conditionals checking for a user will evaluate to false on the first
render (but not later renders, by which it's too late!).

The way to circumvent this is to place the user in localStorage and
initialize Context state with this value. We can then update both
localStorage and state appropriately using Firebase's onAuthStateChanged
listener. Fun!