Fix blank profile immediately after FB reg
There was an issue with the avatar image and user's name not appearing
on the profile page immediately after registration. This is because
the displayName and photoURL are not added to the user until after the
user has been created.

So, RedirectUser fires because the FB user object
updates and redirects to the profile page and does not wait for
updateProfile to complete. onAuthStateChanged only fires when the user
status changes; it does not fire for profile updates. So, the user will
only update on a page refresh.

So, I've added an extra conditional to RedirectUser. If the route that
triggered the redirect was the planform and there is no displayName, it
cancels the redirect. This allows planform page to wait until
updateProfile completes before navigating to the profile page.